add check for cast double->float
Script scoring might result in overflow, for example if a user
decides to use the timestamp as a boost. Therefore, check
if cast cuases a huge precision loss. Note that this does
not always prevent casting issues, since SearchScript has two
methods runAsFloat() and runAsDouble(). Depending on the implementation,
the cast check might not help, see for example AbstractFloatSearchScript and
MvelSearchScript.

+++ src/main/java/org/elasticsearch/common/lucene/search/function/BoostScoreFunction.java
@@ -33,7 +33,6 @@ public class BoostScoreFunction implements ScoreFunction {
         this.boost = boost;
     }
 
-
     public float getBoost() {
         return boost;
     }
@@ -44,8 +43,8 @@ public class BoostScoreFunction implements ScoreFunction {
     }
 
     @Override
-    public float score(int docId, float subQueryScore) {
-        return subQueryScore * boost;
+    public double score(int docId, double subQueryScore) {
+        return subQueryScore * (double) boost;
     }
 
     @Override
@@ -68,12 +67,15 @@ public class BoostScoreFunction implements ScoreFunction {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o)
+            return true;
+        if (o == null || getClass() != o.getClass())
+            return false;
 
         BoostScoreFunction that = (BoostScoreFunction) o;
 
-        if (Float.compare(that.boost, boost) != 0) return false;
+        if (Float.compare(that.boost, boost) != 0)
+            return false;
 
         return true;
     }
